requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
audioread==3.0.1
backoff==2.2.1
bcp47==0.1.0
beautifulsoup4==4.12.3
bert-score==0.3.13
blobfile==3.0.0
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi==1.17.1
chardet==3.0.4
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click @ file:///work/perseverance-python-buildout/croot/click_1698845879718/work
cohere==5.12.0
colorama==0.4.6
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
datasets==3.0.1
decorator==5.1.1
deep-translator==1.11.4
dill==0.3.8
docker-pycreds @ file:///Users/ktietz/demo/mc3/conda-bld/docker-pycreds_1630654474270/work
editdistance==0.8.1
einops==0.8.0
-e git+ssh://git@github.com/dmort27/epitran.git@61a77866c5840337f7d176d953d34210e3ce7a35#egg=epitran
evaluate==0.4.3
fastavro==1.9.7
fasttext==0.9.3
filelock==3.16.1
flash_attn==2.7.4.post1
fonttools==4.55.2
frozenlist==1.4.1
fsspec==2024.6.1
funasr==1.2.6
g2pK==0.9.4
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///croot/gitpython_1720455034730/work
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-trans-new==1.1.9
googletrans==3.0.0
grpcio==1.67.0
h11==0.14.0
h2==3.2.0
hf_transfer==0.1.9
hpack==3.0.0
hstspreload==2024.10.1
httpcore==1.0.7
httpx==0.28.0
httpx-sse==0.4.0
huggingface-hub==0.29.3
hydra-core==1.3.2
hyperframe==5.2.0
HyperPyYAML==1.2.2
idna==2.10
iso639-lang==2.4.2
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
jiwer==3.0.4
jmespath==0.10.0
joblib==1.4.2
JPype1==1.5.0
kagglehub==0.3.6
kaldiio==2.18.1
kiwisolver==1.4.7
konlpy==0.6.0
langid==1.1.6
lazy_loader==0.4
libretranslatepy==2.1.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-bolts==0.7.0
lightning-utilities==0.11.7
llvmlite @ file:///croot/llvmlite_1720455102465/work
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.3
melo==1.1.0
mkl-service==2.4.0
mkl_fft @ file:///croot/mkl_fft_1725370245198/work
mkl_random @ file:///croot/mkl_random_1725370241878/work
modelscope==1.23.2
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
networkx @ file:///croot/networkx_1737039604450/work
nltk==3.9.1
numba @ file:///croot/numba_1720540961817/work
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
oss2==2.19.1
packaging==24.1
pandas==2.2.3
panphon==0.21.2
parameterized==0.9.0
peft==0.13.0
pillow @ file:///croot/pillow_1738010226202/work
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pooch==1.8.2
portalocker==2.10.1
praat-textgrids==1.4.0
protobuf==4.25.3
psutil @ file:///work/perseverance-python-buildout/croot/psutil_1698863411559/work
pyarrow==17.0.0
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
pynndescent==0.5.13
pyparsing==3.2.0
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-mecab-ko==1.3.7
python-mecab-ko-dic==2.1.1.post2
pytorch-lightning==2.4.0
pytorch-wpe==0.0.1
pytz==2024.2
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
RapidFuzz==3.10.0
regex==2024.9.11
requests @ file:///croot/requests_1721410876868/work
rfc3986==1.5.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk @ file:///work/perseverance-python-buildout/croot/sentry-sdk_1698901418049/work
setproctitle @ file:///work/perseverance-python-buildout/croot/setproctitle_1698901493537/work
setuptools==75.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
speechbrain==1.0.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.4.1
torch-complex==0.4.4
torchao==0.8.0
torchaudio==2.5.1
torchmetrics==1.4.2
torchtune==0.5.0
torchvision==0.20.1
tqdm==4.66.5
transformers==4.49.0
transformers-stream-generator==0.0.5
translate==3.6.1
triton==3.0.0
types-requests==2.32.0.20241016
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2024.2
umap-learn==0.5.7
unicodecsv==0.14.1
urllib3 @ file:///croot/urllib3_1727769808118/work
wandb==0.17.8
Werkzeug==3.0.6
wheel==0.44.0
xxhash==3.5.0
yarl==1.13.1
